dev rev 31: -> log only active quest variables, -> Create Shop log more accurate, -> Check for some rewards (direct ones) for Quest Hacks (Santa for now), -> Santa Quest (we take the stolen santa suit, give items and delete quest so we can have more items), -> new function "DelInactiveQuest" used for Santa Rewards but later for non essential quests, -> file "binary/pthreadGC.dll" added in dev SVN, -> update of .bat "database/database_installer.bat" in order to use chest_data.sql, -> update of "database/quest_data.sql", santa Event and some stuff, -> added a trap for the GG packet (0x808) in charserver (just returns true for now),
Special Santa thing:
The give items are in quest_data, from 8000 to 8003 so people can change items / family ID directly there so they can put other clothes or other santa look.
The rewards are loggued and checked in case of attempt of quest hack (numbers of stolen suit and if quest actually exists).